♻️ Refactor 4.02_Extract smart add guide
